Melhorando as validações do backend

+++ src/app/controllers/ContactController.js
@@ -1,4 +1,5 @@
 const ContactsRepository = require('../repositories/ContactsRepository');
+const isValidUUID = require('../utils/isValidUUID');
 
 class ContactController {
   async index(request, response) {
@@ -13,10 +14,13 @@ class ContactController {
     // Obter um registro
     const { id } = request.params;
 
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact ID' });
+    }
+
     const contact = await ContactsRepository.findById(id);
 
     if (!contact) {
-      // 404: Not Found
       return response.status(404).json({ error: 'Contact not found' });
     }
 
@@ -33,14 +37,19 @@ class ContactController {
       return response.status(400).json({ error: 'Name is required' });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category_ID' });
+    }
 
-    if (contactExists) {
-      return response.status(400).json({ error: 'This e-mail is already in use' });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      if (contactExists) {
+        return response.status(400).json({ error: 'This e-mail is already in use' });
+      }
     }
 
     const contact = await ContactsRepository.create({
-      name, email, phone, category_id,
+      name, email: email || null, phone, category_id: category_id || null,
     });
     response.json(contact);
   }
@@ -48,17 +57,28 @@ class ContactController {
   async update(request, response) {
     // Editar um registro
     const { id } = request.params;
+
     const {
       name, email, phone, category_id,
     } = request.body;
 
-    const contactExists = await ContactsRepository.findById(id);
-    if (!contactExists) {
-      return response.status(404).json({ error: 'Contact not found' });
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact ID' });
+    }
+
+    if (category_id && !isValidUUID(category_id)) {
+      return response.status(400).json({ error: 'Invalid category ID' });
     }
 
     if (!name) {
       return response.status(400).json({ error: 'Name is required' });
+    }
+
+    if (email) {
+      const contactExists = await ContactsRepository.findById(id);
+      if (!contactExists) {
+        return response.status(404).json({ error: 'Contact not found' });
+      }
     }
 
     const contactByEmail = await ContactsRepository.findByEmail(email);
@@ -67,7 +87,7 @@ class ContactController {
     }
 
     const contact = await ContactsRepository.update(id, {
-      name, email, phone, category_id,
+      name, email: email || null, phone, category_id: category_id || null,
     });
 
     response.json(contact);
@@ -77,15 +97,12 @@ class ContactController {
     // Deletar um registro
     const { id } = request.params;
 
-    // const contact = await ContactsRepository.findById(id);
-
-    // if (!contact) {
-    //   // 404: Not Found
-    //   return response.status(404).json({ error: 'User not found' });
-    // }
+    if (!isValidUUID(id)) {
+      return response.status(400).json({ error: 'Invalid contact ID' });
+    }
 
     await ContactsRepository.delete(id);
-    // 204: No Content
+
     response.sendStatus(204);
   }
 }
